+++ src/video/SDLOffScreenSurface.cc
@@ -42,7 +42,7 @@ void SDLOffScreenSurface::saveScreenshot(const std::string& filename)
 
 void SDLOffScreenSurface::clearScreen()
 {
-	memset(surface->pixels, 0, uint32_t(surface->pitch) * surface->h);
+	memset(surface->pixels, 0, size_t(surface->pitch) * surface->h);
 }
 
 } // namespace openmsx
